% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ar1}
\alias{ar1}
\title{Function to compute the AR(1) correlation coefficient for a given data series}
\usage{
ar1(data)
}
\arguments{
\item{data}{data frame containing daily discharge data}
}
\value{
ar1 AR(1) correlation coefficient
}
\description{
This function accepts a data frame containing daily streamflow data and returns the AR(1)
correlation coefficient
}
\examples{
qfiletempf<-sampleData
ar1(qfiletempf)
}

