% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bfi.R
\name{calc_bfi}
\alias{calc_bfi}
\title{Function to return the base flow index for a given data frame}
\usage{
calc_bfi(x)
}
\arguments{
\item{x}{A numeric vector of consecutive daily mean discharge values}
}
\value{
calc_bfi numeric value of the base flow index for the given data frame
}
\description{
This function accepts a vector of daily mean discharge values and 
calculates the base flow index of the daily flow values for the entire record
}
\details{
The mean for a 7 day right-aligned moving window is calculated for the supplied flow vector. The baseflow index is calculated as the minimum 7day average flow divided by the mean flow
}
\examples{
x<-sampleData$discharge
calc_bfi(x)
}
