% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitDurationLow.R
\name{hitDurationLow}
\alias{hitDurationLow}
\title{Indices describing duration of low flow events.}
\usage{
hitDurationLow(x, yearType = "water", digits = 3, pref = "mean", ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 24 indices used to describe the duration of low flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item dl1 Annual minimum daily flow. Compute the minimum 1-day average flow for each year. DL1 is the mean 
(or median-Use Preference option) of these values.
\item dl2 Annual minimum of 3-day moving average flow. Compute the minimum of a 3-day moving average flow for 
each year. DL2 is the mean (or median-Use Preference option) of these values. 
\item dl3 Annual minimum of 7-day moving average flows. Compute the minimum of a 7-day moving average flow for 
each year. DL3 is the mean (or median-Use Preference option) of these values.
\item dl4 Annual minimum of 30-day moving average flows. Compute the minimum of a 30-day moving average flow for 
each year. DH4 is the mean (or median-Use Preference option) of these values.
\item dl5 Annual minimum of 90-day moving average flows. Compute the minimum of a 90-day moving average flow for 
each year. DH5 is the mean (or median-Use Preference option) of these values.
\item dl6 Variability of annual minimum daily average flow. Compute the standard deviation for the minimum daily 
average flow. DL6 is 100 times the standard deviation divided by the mean.
\item dl7 Variability of annual minimum of 3-day moving average flows. Compute the standard deviation for the 
minimum 3-day moving averages. DL7 is 100 times the standard deviation divided by the mean.
\item dl8 Variability of annual minimum of 7-day moving average flows. Compute the standard deviation for the 
minimum 7-day moving averages. DL8 is 100 times the standard deviation divided by the mean.
\item dl9 Variability of annual minimum of 30-day moving average flows. Compute the standard deviation for the 
minimum 30-day moving averages. DL9 is 100 times the standard deviation divided by the mean.
\item dl10 Variability of annual minimum of 90-day moving average flows. Compute the standard deviation for the 
minimum 90-day moving averages. DH10 is 100 times the standard deviation divided by the mean.
\item dl11 Annual minimum daily flow divided by the median for the entire record. Compute the minimum daily flow 
for each year. DL11 is the mean of these values divided by the median for the entire record.
\item dl12 Annual minimum of 7-day moving average flows divided by the median for the entire record. Compute the 
minimum of a 7-day moving average flow for each year. DL12 is the mean of these values divided by the median 
for the entire record.
\item dl13 Annual minimum of 30-day moving average flows divided by the median for the entire record. Compute the 
minimum of a 30-day moving average flow for each year. DL13 is the mean of these values divided by the median 
for the entire record.
\item dl14 Low exceedence flows. Compute the 75-percent exceedence value for the entire flow record. DL14 is 
the exceedence value divided by the median for the entire record.
\item dl15 Low exceedence flows. Compute the 90-percent exceedence value for the entire flow record. DL15 is the 
exceedence value divided by the median for the entire record.
\item dl16 Low flow pulse duration. Compute the average pulse duration for each year for flow events below a 
threshold equal to the 25th percentile value for the entire flow record. DL16 is the median of the yearly 
average durations.
\item dl17 Variability in low pulse duration. Compute the standard deviation for the yearly average low pulse durations. 
DL17 is 100 times the standard deviation divided by the mean of the yearly average low pulse durations.
\item dl18 Number of zero-flow days. Count the number of zero-flow days for the entire flow record. DL18 is the 
mean (or median-Use Preference option) annual number of zero flow days.
\item dl19 Variability in the number of zero-flow days. Compute the standard deviation for the annual number of 
zero-flow days. DL19 is 100 times the standard deviation divided by the mean annual number of zero-flow days.
\item dl20 Number of zero-flow months. While computing the mean monthly flow values, count the number of months 
in which there was no flow over the entire flow record.
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
hitDurationLow(x=x,yearType=yearType)

}
