% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitFrequencyHigh.R
\name{hitFrequencyHigh}
\alias{hitFrequencyHigh}
\title{Indices describing frequency of high flow events.}
\usage{
hitFrequencyHigh(x, yearType = "water", digits = 3, pref = "mean",
  floodThreshold = NULL, ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{floodThreshold}{Numeric flood threshold as the flow equivalent for a flood recurrence of 1.67 years}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 11 indices used to describe the frequency of high flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item fh1 High flood pulse count. Compute the average number of flow events with flows above a threshold equal to 
the 75th percentile value for the entire flow record. FH1 is the average (or median-Use Preference option) 
number of events.
\item fh2 Variability in high pulse count. Compute the standard deviation in the annual pulse counts for FH1. FH2 
is 100 times the standard deviation divided by the mean pulse count (number of events/year-spatial).  
\item fh3 High flood pulse count. Compute the average number of days per year that the flow is above a threshold equal 
to three times the median flow for the entire record. FH3 is the mean (or median-Use Preference option) of the 
annual number of days for all years.
\item fh4 High flood pulse count. Compute the average number of days per year that the flow is above a threshold 
equal to seven times the median flow for the entire record. FH4 is the mean (or median - Use Preference option) 
of the annual number of days for all years.
\item fh5 Flood frequency. Compute the average number of flow events with flows above a threshold equal to the 
median flow value for the entire flow record. FH5 is the average (or median - Use Preference option) number 
of events.
\item fh6 Flood frequency. Compute the average number of flow events with flows above a threshold equal to three 
times the median flow value for the entire flow record. FH6 is the average (or median-Use Preference option) 
number of events.
\item fh7 Flood frequency. Compute the average number of flow events with flows above a threshold equal to 
seven times the median flow value for the entire flow record. FH7 is the average (or median-Use Preference option) 
number of events.
\item fh8 Flood frequency. Compute the average number of flow events with flows above a threshold equal to 
25-percent exceedence value for the entire flow record. FH8 is the average (or median-Use Preference option) 
number of events.
\item fh9 Flood frequency. Compute the average number of flow events with flows above a threshold equal to 
75-percent exceedence value for the entire flow record. FH9 is the average (or median-Use Preference option) 
number of events.
\item fh10 Flood frequency. Compute the average number of flow events with flows above a threshold equal to 
median of the annual minima for the entire flow record. FH10 is the average (or median-Use Preference option) 
number of events.
\item fh11 Flood frequency. Compute the average number of flow events with flows above a threshold equal to flow 
corresponding to a 1.67-year recurrence interval. FH11 is the average (or median-Use Preference option) number 
of events.
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
floodThreshold = 1158
hitFrequencyHigh(x=x,yearType=yearType,floodThreshold = 1158)
}
