% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitMagHigh.R
\name{hitMagHigh}
\alias{hitMagHigh}
\title{Indices describing magnitude of the peak flow condition.}
\usage{
hitMagHigh(x, yearType = "water", digits = 3, drainArea = NULL,
  pref = "mean", ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{drainArea}{A numeric specifying the drainage area. Only required for mh20 statistic. Typically squiare miles, see details.}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 27 indices used to describe the magnitude of the peak flow condition. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item mh1_12  Requires pref argument to be either "mean" or "median" specifying monthly aggregation function. 
Default is "mean". Means (or medians - use preference option) of maximum daily flow values for each month. 
For example, mh1 is the mean of all January maximum flow values over the entire record. 
\item mh13 variability (coefficient of variation) across maximum monthly flow values. Compute the mean 
and standard deviation for the maximum monthly flows over the entire flow record. MH13 is the standard deviation 
times 100 divided by the mean maximum monthly flow for all years.
\item mh14 median of annual maximum flows. Compute the annual maximum flows from monthly maximum flows. 
Compute the ratio of annual maximum flow to median annual flow for each year. MH14 is the median of these ratios.
\item mh15_17 MH15; High flow discharge index. Compute the 1-percent exceedence value for the entire data record. MH15 is 
the 1-percent exceedence value divided by the median flow for the entire record. 
MH16; Compute the 10-percent exceedence value for the entire data record. MH16 is the 10-percent exceedence 
value divided by the median flow for the entire record. 
MH17; Compute the 25-percent exceedence value for the entire data record. MH17 is the 25-percent exceedence 
value divided by the median flow for the entire record. 
\item mh18 variability across annual maximum flows. Compute the logs (log10) of the maximum annual flows. 
Find the standard deviation and mean for these values. MH18 is the standard deviation times 100 divided by the 
mean.
\item mh19 the skewness in annual maximum flows (dimensionless-spatial). Use the equation:
MH19   numerator =   N2 ? sum(qm3)-3N ? sum(qm) ? sum(qm2) + 2 ? (sum(qm))3  
denominator = N ? (N-1) ? (N-2) ? stddev3  
Where:  N = Number of years
qm = Log10 (annual maximum flows)
stddev = Standard deviation of the annual maximum flows
\item mh20 specific mean annual maximum flow. MH20 is the mean (or median-Use Preference option) of 
the annual maximum flows divided by the drainage area (cubic feet per second/square mile-temporal).
\item mh21_27 high flow volume indices. Compute the average volume for flow events above a threshold.
Thresholds are equal to the median flow for the entire record for mh21, 3 times the median flow for the 
entire record for mh22,and 7 times the median flow for the entire record for mh23. 
Thresholds are equal to the median flow for the entire record for mh24, 3 times the median flow for the 
entire record for mh25, 7 times the median flow for the entire record for mh26, 
and the 75th percentile for the entire record for mh27. 
MH21 through 23 are the average volumes divided by the median flow for the entire 
record. MH24 through 27 are the average peak flows divided by the median flow for the 
entire record.
}
}
\examples{
x <- sampleData[c("date","discharge")]
drainArea <- 50
yearType = "water"
hitMagHigh(x=x,yearType=yearType,drainArea=drainArea)

}
