% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitMagLow.R
\name{hitMagLow}
\alias{hitMagLow}
\title{Indices describing magnitude of the low flow condition.}
\usage{
hitMagLow(x, yearType = "water", digits = 3, drainArea = NULL,
  pref = "mean", ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{drainArea}{A numeric specifying the drainage area. Only required for ml22 statistic. Typically squiare miles, see details.}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 27 indices used to describe the magnitude of the peak flow condition. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item mh1_12  Requires pref argument to be either "mean" or "median" specifying monthly aggregation function. 
Default is "mean". Means (or medians - use preference option) of maximum daily flow values for each month. 
For example, mh1 is the mean of all January maximum flow values over the entire record. 
\item mh13 variability (coefficient of variation) across minimum monthly flow values. Compute the mean 
and standard deviation for the maximum monthly flows over the entire flow record. MH13 is the standard deviation 
times 100 divided by the mean maximum monthly flow for all years.
\item ml14 Mean of annual minimum annual flows.  ML14 is the mean of the ratios of minimum annual flows to the median 
flow for each year.
\item ml15 Low flow index. ML15 is the mean (or median-Use Preference option) of the ratios of minimum annual flows to 
the mean flow for each year.
\item ml16 Median of annual minimum flows.  ML16 is the median of the ratios of minimum annual flows to the median 
flow for each year. 
\item ml17 Baseflow 1. Compute the mean annual flows. Compute the minimum of a 7-day moving average flow 
for each year and divide them by the mean annual flow for that year. ML17 is the mean (or median-Use 
Preference option) of those ratios. 
\item ml18 Variability in baseflow 1. Compute the standard deviation for the ratios of minimum 7-day 
moving average flows to mean annual flows for each year.  ML18 is the standard deviation times 100 divided by 
the mean of the ratios. 
\item ml19 Baseflow 2. Compute the ratios of the minimum annual flow to mean annual flow for 
each year. ML19 is the mean (or median-Use Preference option) of these ratios times 100.
\item ml20 Baseflow 3. Divide the daily flow record into 5-day blocks. Find the minimum flow for 
each block. Assign the minimum flow as a base flow for that block if 90 percent of that minimum flow is less 
than the minimum flows for the blocks on either side. Otherwise, set it to zero. Fill in the zero values 
using linear interpolation. Compute the total flow for the entire record and the total base flow for the 
entire record. ML20 is the ratio of total base flow to total flow.
\item ml21 Variability across annual minimum flows. Compute the mean and standard deviation for the 
annual minimum flows. ML21 is the standard deviation times 100 divided by the mean. 
\item ml22 Specific mean annual minimum flow.  ML22 is the mean (or median-Use Preference option) of the 
annual minimum flows divided by the drainage area.
}
}
\examples{
x <- sampleData[c("date","discharge")]
drainArea <- 50
yearType = "water"
hitMagLow(x=x,yearType=yearType,drainArea=drainArea)

}
