% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_magnifSeven.r
\name{calc_magnifSeven}
\alias{calc_magnifSeven}
\title{Function to return the magnificent seven statistics for a given data series}
\usage{
calc_magnifSeven(x, yearType = "water", wyMonth = 10L, digits = 3)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{wyMonth}{The month of the year in which the water year starts.}

\item{digits}{A numeric. Number of digits to round indice values}
}
\value{
data.frame of calculated statistics
}
\description{
Function to return the magnificent seven statistics for a given data series
}
\details{
This is a function to compute the 7 statistics of daily streamflow
used by Archfield et al., under revision (June 2013).
}
\examples{
x<-sampleData[c("date","discharge")]
magSeven <- calc_magnifSeven(x)
}
