% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitTimingAverage.R
\name{hitTimingAverage}
\alias{hitTimingAverage}
\title{Indices describing timing of average flow events.}
\usage{
hitTimingAverage(x, yearType = "water", digits = 3, pref = "mean",
  floodThreshold = NULL, ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{floodThreshold}{Numeric flood threshold as the flow equivalent for a flood recurrence of 1.67 years}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 3 indices used to describe the timing of average flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item  ta1; Constancy. Constancy is computed via the formulation of Colwell (see example in
Colwell, 1974). A matrix of values is compiled where the columns are 11 flow
categories and the rows are 365 days of the year (no leap years) defined as either calendar or water year. February 29th is removed on leap years. The cell
values are the number of times that a flow falls into a category on each day. The
categories are listed below. The row totals, column totals, and grand total are computed. Using the equations for
Shannon information theory parameters, constancy is computed as 1-(uncertainty with respect to state)/log10(number of state) 
\itemize{
\item log(flow) < .1 x log(mean flow)
\item .1 x log(mean flow) <= log(flow) < .25 x log(mean flow) 
\item .25 x log(mean flow) <= log(flow) < .5 x log(mean flow) 
\item .5 x log(mean flow) <= log(flow) < .75 x log(mean flow)
\item .75 x log(mean flow) <= log(flow) < 1.0 x log(mean flow)
\item 1.0 x log(mean flow) <= log(flow) < 1.25 x log(mean flow) 
\item 1.25 x log(mean flow) <= log(flow) < 1.5 x log(mean flow)
\item 1.5 x log(mean flow) <= log(flow) < 1.75 x log(mean flow)
\item 1.75 x log(mean flow) <= log(flow) < 2.0 x log(mean flow)
\item 2.0 x log(mean flow) <= log(flow) < 2.25 x log(mean flow) 
\item log(flow) >= 2.25 x log(mean flow)
} 
\item ta2; Predictability. Predictability is computed from the same matrix as constancy (see
example in Colwell, 1974). It is computed as: 1- (uncertainty with respect to interaction of time and state - uncertainty with respect to time)/log10(number of state)
\item ta3; Seasonal predictability of flooding. Divide years up into 2-month periods (that is,
Oct-Nov, Dec-Jan, and so forth). Count the number of flood days (flow events with
flows > 1.67-year flood) in each period over the entire flow record. TA3 is the
maximum number of flood days in any one period divided by the total number of
flood days. 
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
floodThreshold = 1158
hitTimingAverage(x=x,yearType = yearType,floodThreshold =floodThreshold)
}
