% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitTimingHigh.R
\name{hitTimingHigh}
\alias{hitTimingHigh}
\title{Indices describing timing of high flow events.}
\usage{
hitTimingHigh(x, yearType = "water", digits = 3, pref = "mean",
  floodThreshold = NULL, ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{floodThreshold}{Numeric flood threshold as the flow equivalent for a flood recurrence of 1.67 years}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 3 indices used to describe the timing of high flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item  th1 Julian date of annual maximum. Determine the Julian date that the maximum flow occurs for each year. 
Transform the dates to relative values on a circular scale (radians or degrees). Compute the x and y components 
for each year and average them across all years. Compute the mean angle as the arc tangent of y-mean divided by 
x-mean. Transform the resultant angle back to Julian date.
\item th2 Variability in Julian date of annual maxima. Compute the coefficient of variation for the mean x and y 
components and convert to a date.
\item th3 Seasonal predictability of nonflooding. Computed as the maximum proportion of a 365-day year that the 
flow is less than the 1.67-year flood threshold and also occurs in all years. Accumulate nonflood days that 
span all years. TH3 is maximum length of those flood-free periods divided by 365.  
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
floodThreshold = 1158
hitTimingHigh(x=x,yearType=yearType,floodThreshold=floodThreshold)
}
