% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitTimingLow.R
\name{hitTimingLow}
\alias{hitTimingLow}
\title{Indices describing timing of low flow events.}
\usage{
hitTimingLow(x, yearType = "water", digits = 3, pref = "mean",
  floodThreshold = NULL, ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{floodThreshold}{Numeric flood threshold as the flow equivalent for a flood recurrence of 1.67 years}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 3 indices used to describe the timing of high flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item  tl1; Julian date of annual minimum. Determine the Julian date that the minimum flow occurs for each water year. 
Transform the dates to relative values on a circular scale (radians or degrees). Compute the x and y components 
for each year and average them across all years. Compute the mean angle as the arc tangent of y-mean divided by 
x-mean. Transform the resultant angle back to Julian date.
\item tl2 Variability in Julian date of annual minima. Compute the coefficient of variation for the mean x and y 
components and convert to a date.
\item tl3 Seasonal predictability of low flow. Divide years up into 2-month periods (that is, Oct-Nov, Dec-Jan, and 
so forth). Count the number of low flow events (flow events with flows <= 5 year flood threshold) in each period 
over the entire flow record. TL3 is the maximum number of low flow events in any one period divided by the total 
number of low flow events.
\item tl4 Seasonal predictability of non-low flow. Compute the number of days that flow is above the 5-year flood 
threshold as the ratio of number of days to 365 or 366 (leap year) for each year. TL4 is the maximum of the yearly 
ratios.
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
floodThreshold = 1161.38
hitTimingLow(x=x,yearType=yearType,floodThreshold=floodThreshold)
}
