% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcChangeEvents.R
\name{calcChangeEvents}
\alias{calcChangeEvents}
\title{Function to decompose a flow series into flow direction change events}
\usage{
calcChangeEvents(x)
}
\arguments{
\item{x}{A vector of flow values, should be sorted chronologically.}
}
\value{
A dataframe with columns "flow" and "event"
}
\description{
Function to decompose a flow series into flow direction change events
}
\details{
This function accepts a vector of flow values, tests if each value in the flow vector
is greater or less than the preceding value of flow, and classifies the flow vector into events. An event
is defined as when a change in flow from one day to the next changes direction, e.g. rising limb to falling limb. 
If there is no change in flow from one day to the next it is considered part of the preceding event.
}
\examples{
x <- sampleData$discharge
calcChangeEvents(x)
}

