% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEvents.R
\name{calcEvents}
\alias{calcEvents}
\title{Function to decompose a flow series into high or low flow events.}
\usage{
calcEvents(x, threshold, type = "high")
}
\arguments{
\item{x}{A numeric vector of flow values}

\item{threshold}{Numeric threshold value defining event}

\item{type}{character indicating type of event. High flags events above threshold, low flags events below threshold.}
}
\value{
A dataframe with columns "flow" and "event"
}
\description{
This function accepts a vector of flow values, tests if each value in the flow vector
is above or below a defined threshold, and classifies the flow vector into events. An event
is defined as consecutive values above or below the given threshold.
}
\examples{
x <- sampleData$discharge
threshold <- median(x,na.rm=TRUE)
calcEvents(x,threshold)
}
