% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highFlowEvents.R
\name{calcHighEvents}
\alias{calcHighEvents}
\title{Function to decompose a flow series into high flow events defined as flows above a given threshold.}
\usage{
calcHighEvents(x, threshold)
}
\arguments{
\item{flow}{A vector of flow values}
}
\value{
A dataframe with columns "flow" and "event"
}
\description{
This function accepts a vector of flow values, tests if each value in the flow vector
is above a defined threshold, and classifies the flow vector into events. An event
is defined as consecutive entries above the given threshold.
}
\examples{
x <- sampleData$flow
threshold <- median(x,na.rm=TRUE)
highFlowEvents(x,threshold)
}

