% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Function to return the coefficient of variation for a given data series}
\usage{
cv(x)
}
\arguments{
\item{x}{numeric vector of flow values}
}
\value{
numeric coefficient of variation for the given data frame
}
\description{
Function to return the coefficient of variation for a given data series
}
\details{
This functions accepts a numeric vector of flow values and calculates the coefficient of variation, defined as the mean/standard deviation
}
\examples{
cv(sampleData$discharge)
}
