% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dl14}
\alias{dl14}
\title{Function to return the DL14 hydrologic indicator statistic for a given data frame}
\usage{
dl14(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
dl14 numeric containing DL14 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates
DL14; Low exceedence flows. Compute the 75-percent exceedence value for the entire flow record. DL14 is
the exceedence value divided by the median for the entire record (dimensionless-spatial).
}
\examples{
qfiletempf<-sampleData
dl14(qfiletempf)
}

