% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dl4}
\alias{dl4}
\title{Function to return the DL4 hydrologic indicator statistic for a given data frame}
\usage{
dl4(qfiletempf, pref = "mean")
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}

\item{pref}{string containing a "mean" or "median" preference}
}
\value{
dl4 numeric containing DL4 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates
DL4; Annual minimum of 30-day moving average flow. Compute the minimum of a 30-day moving average flow
for each year. DL4 is the mean (or median-Use Preference option) of these values (cubic feet per second-temporal).
}
\examples{
qfiletempf<-sampleData
dl4(qfiletempf)
}

