% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dl8}
\alias{dl8}
\title{Function to return the DL8 hydrologic indicator statistic for a given data frame}
\usage{
dl8(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
dl8 numeric containing DL8 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates
DL8; Variability of annual minimum of 7-day moving average flow. Compute the standard deviation for the
minimum 7-day moving averages. DL8 is 100 times the standard deviation divided by the mean (percent-spatial).
}
\examples{
qfiletempf<-sampleData
dl8(qfiletempf)
}

