% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDuration.R
\name{eventDuration}
\alias{eventDuration}
\title{Function to decompose a flow series into high flow events defined as flows above a given threshold.}
\usage{
eventDuration(x, threshold, average = TRUE, type = "high", pref = "mean")
}
\arguments{
\item{threshold}{Numeric threshold value for high flow events}

\item{average}{Logical. If TRUE returns only the average event duration. If FALSE returns the duration for each individual event}

\item{flow}{A vector of numeric flow values}
}
\value{
A numeric of the mean event duration in units of number of rows if average = TRUE. Otherwise, returns a datafarme of durations for each event.
}
\description{
This function accepts a vector of flow values, tests if each value in the flow vector
is above a defined threshold, classifies the flow vector into events, and returns the 
average event duration. An event
is defined as consecutive entries above the given threshold.
}
\examples{
x <- sampleData$discharge
threshold <- median(x,na.rm=TRUE)
eventDuration(x,threshold)
}

