% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fh4}
\alias{fh4}
\title{Function to return the FH4 hydrologic indicator statistic for a given data frame}
\usage{
fh4(qfiletempf, pref = "mean")
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}

\item{pref}{string containing a "mean" or "median" preference}
}
\value{
fh4 numeric value of high flood pulse count for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates
FH4; High flood pulse count. Compute the average number of days per year that the flow is above a threshold
equal to seven times the median flow for the entire record. FH4 is the mean (or median - Use Preference option)
of the annual number of days for all years (number of days/year-temporal).
}
\examples{
qfiletempf<-sampleData
fh4(qfiletempf)
}

