% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh5.R
\name{fh5}
\alias{fh5}
\title{Function to return the FH5 hydrologic indicator statistic for a given data frame}
\usage{
fh5(qfiletempf, pref = "mean")
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}

\item{pref}{string containing a "mean" or "median" preference}
}
\value{
fh5 numeric value of FH5 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates 
FH5; Flood frequency. Compute the average number of flow events with flows above a threshold equal to the 
median flow value for the entire flow record. FH5 is the average (or median - Use Preference option) number 
of events (number of events/year-temporal).
}
\examples{
qfiletempf<-sampleData
fh5(qfiletempf)
}

