% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eventDuration.R
\name{find_eventDuration}
\alias{find_eventDuration}
\title{Function to decompose a flow series into high flow events defined as flows above a given threshold.}
\usage{
find_eventDuration(x, threshold, aggType = "average", type = "high",
  pref = "mean", trim = FALSE)
}
\arguments{
\item{x}{A vector of numeric flow values}

\item{threshold}{Numeric threshold value for high flow events}

\item{aggType}{Character vector indicating type of aggregation to use, if any. Choices are "average", "min", "max" or "none". If "average", mean or median of events is returned depending on "pref" setting If "none", returns the duration for each individual event}

\item{type}{Character of either "high" or "low" indicating if events are defined as flows below or above the given threshold, respectively}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{trim}{Logical. Events that start or end at the beginning or end of the record are dropped if \code{TRUE} because an accurate duration can't be calculated if teh start or end time is unknown.}
}
\value{
A numeric of the mean event duration in units of number of rows if average = TRUE. Otherwise, returns a datafarme of durations for each event.
}
\description{
This function accepts a vector of flow values, tests if each value in the flow vector
is above or below a defined threshold, classifies the flow vector into events, and returns the 
average event duration or all event durations if average = FALSE. An event
is defined as consecutive entries above the given threshold.
}
\examples{
x <- sampleData$discharge
threshold <- median(x,na.rm=TRUE)
find_eventDuration(x,threshold)
}
