% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findrank.R
\name{findrank}
\alias{findrank}
\title{Function to return the rank of a given value}
\usage{
findrank(n, p)
}
\arguments{
\item{n}{number of values in a data set}

\item{p}{percentile desired}
}
\value{
findrank numeric giving the position in the data set of the percentile
}
\description{
This function accepts the number of values in a data set and the desired percentile and then calculates the rank
}
\examples{
n<-365
p<-0.05
findrank(n, p)
}

