% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_perc.R
\name{flow_perc}
\alias{flow_perc}
\title{Function to return the requested flow percentiles for a given data series}
\usage{
flow_perc(data, probs = c(0.1, 0.25, 0.5, 0.75, 0.9, 0.15))
}
\arguments{
\item{data}{data frame containing value data for the chosen timeseries}

\item{probs}{vector containing requested percentile value(s)}
}
\value{
obs_percentiles requested flow percentiles for the given data frame
}
\description{
This function accepts a data frame containing daily data and a list of desired percentiles and
returns a list of the requested percentiles
}
\examples{
qfiletempf<-sampleData
flow_perc(qfiletempf,probs=c(.1,.25,.5,.75))
}

