% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waterYear.R
\name{get_waterYear}
\alias{get_waterYear}
\title{Function to return the water year for a given date}
\usage{
get_waterYear(x, wyMonth = 10L, numeric = TRUE)
}
\arguments{
\item{x}{A date vector}

\item{wyMonth}{A value of class integer}

\item{numeric}{Logical. Return values are numeric vector or vector of ordered factors}
}
\value{
A vector of numeric water years for each date if \code{numeric = TRUE} otherwise an ordered factor of water years if \code{numeric = FALSE}
}
\description{
Function to return the water year for a given date
}
\examples{
get_waterYear(sampleData$date)

}
