% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waterYearDay.R
\name{get_waterYearDay}
\alias{get_waterYearDay}
\title{Day of water year}
\usage{
get_waterYearDay(x, wyMonth = 10L)
}
\arguments{
\item{x}{A vector of class date.}

\item{wyMonth}{A numeric. The month of the year in which the water year starts 
(1=January, 12=December). The water year begins on the first day of wyMonth.}
}
\value{
A numeric vector of day of water year
}
\description{
Given a vector of dates, calculates day of water year accounting for leap years.
}
\examples{
x <- seq(from=as.Date("2010-01-01"),to=as.Date("2016-01-01"),by="1 days")
year_day <- get_waterYearDay(x, 2L)
}
