% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataLocal.R
\name{getDataLocal}
\alias{getDataLocal}
\title{Function to return flow data from a locally stored file}
\usage{
getDataLocal(dataPath, startDt = "", endDt = "", sepChar = ",")
}
\arguments{
\item{dataPath}{path to directory containing data files}

\item{startDt}{beginning water year, will be translated to 10/01}

\item{endDt}{ending water year, will be translated to 09/30}

\item{sepChar}{string containing the datafile separator, default is comma}
}
\value{
dataOut data frame containing requested flow data for the stations
}
\description{
This function accepts a list of sites, start and end date and returns a data frame containing requested flow data
}
\examples{
dataPath <- system.file("extdata", package="EflowStats")
dataPath <- paste(dataPath, "modeled", sep="/")
startdate <- "2009"
enddate <- "2013"
localData <- getDataLocal(dataPath,startDt=startdate,endDt=enddate)
}

