% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataUSGS.R
\name{getDataUSGS}
\alias{getDataUSGS}
\title{Function to return flow data for a of USGS site and start and end date}
\usage{
getDataUSGS(sites, startdate, enddate)
}
\arguments{
\item{sites}{USGS station ids}

\item{startdate}{beginning water year, will be translated to 10/01}

\item{enddate}{ending water year, will be translated to 09/30}
}
\value{
dataOut data frame containing requested flow data for the stations
}
\description{
This function accepts a list of sites, start and end date and returns a data frame containing requested flow data
}
\examples{
sites <- c("02177000", "02178400")
startdate <- "2009"
enddate <- "2013"
\dontrun{
data <- getDataUSGS(sites,startdate,enddate)
}
}

