% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/peakflow.R
\name{getPeakData}
\alias{getPeakData}
\title{Function to return data from the NWISWeb peak values service}
\usage{
getPeakData(sites)
}
\arguments{
\item{sites}{string containing the site for the retrieval}
}
\value{
peakValues a data frame containing columns 'Date' and 'Discharge'
}
\description{
This function accepts a sites parameter that contains the desired NWIS site. It returns a data frame
containing peak instantaneous annual flow data for the site.
}
\examples{
sites<-"02178400"
\dontrun{
peakData <- getPeakData(sites)
}
}

