% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitDurationHigh.R
\name{hitDurationHigh}
\alias{hitDurationHigh}
\title{Indices describing duration of high flow events.}
\usage{
hitDurationHigh(x, yearType = "water", digits = 3, pref = "mean",
  floodThreshold = NULL, ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{floodThreshold}{Numeric value of flood threshold as the flow equivalent for a flood recurrence of 1.67 years. Can be calculated using the \code{peakThreshold} function.}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 24 indices used to describe the duration of high flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item dh1 Annual maximum daily flow. Compute the maximum of a 1-day moving average flow for each year. dh1 is the 
mean (or median-Use Preference option) of these values.
\item dh2 Annual maximum of 3-day moving average flows. Compute the maximum of a 3-day moving average flow for 
each year. dh2 is the mean (or median-Use Preference option) of these values; 
\item dh3 Annual maximum of 7-day moving average flows. Compute the maximum of a 7-day moving average flow for 
each year. dh3 is the mean (or median-Use Preference option) of these values.
\item dh4 Annual maximum of 30-day moving average flows. Compute the maximum of a 30-day moving average flow for 
each year. dh4 is the mean (or median-Use Preference option) of these values.
\item dh5 Annual maximum of 90-day moving average flows. Compute the maximum of a 90-day moving average flow for 
each year. dh5 is the mean (or median-Use Preference option) of these values.
\item dh6 Variability of annual maximum daily flows. Compute the standard deviation for the maximum 1-day 
moving averages. dh6 is 100 times the standard deviation divided by the mean.
\item dh7 Variability of annual maximum of 3-day moving average flows. Compute the standard deviation for the 
maximum 3-day moving averages. dh7 is 100 times the standard deviation divided by the mean.
\item dh8 Variability of annual maximum of 7-day moving average flows. Compute the standard deviation for the 
maximum 7-day moving averages. dh8 is 100 times the standard deviation divided by the mean.
\item dh9 Variability of annual maximum of 30-day moving average flows. Compute the standard deviation for the 
maximum 30-day moving averages. dh9 is 100 times the standard deviation divided by the mean.
\item dh10 Variability of annual maximum of 90-day moving average flows. Compute the standard deviation for the 
maximum 90-day moving averages. dh10 is 100 times the standard deviation divided by the mean.
\item dh11 Annual maximum of 1-day moving average flows divided by the median for the entire record. Compute the 
maximum of a 1-day moving average flow for each year. dh11 is the mean of these values divided by the median 
for the entire record.
\item dh12 Annual maximum of 7-day moving average flows divided by the median for the entire record. Compute the 
maximum of a 7-day moving average flow for each year. dh12 is the mean of these values divided by the median 
for the entire record.
\item dh13 Annual maximum of 30-day moving average flows divided by the median for the entire record. Compute the 
maximum of a 30-day moving average flow for each year. dh13 is the mean of these values divided by the median 
for the entire record.
\item dh14 Flood duration. Compute the mean of the mean monthly flow values. Find the 95th percentile for the 
mean monthly flows. dh14 is the 95th percentile value divided by the mean of the monthly means.
\item dh15 High flow pulse duration. Compute the average duration for flow events with flows above a threshold equal 
to the 75th percentile value for each year in the flow record. dh15 is the median of the yearly average durations.
\item dh16 Variability in high flow pulse duration. Compute the standard deviation for the yearly average high pulse 
durations. dh16 is 100 times the standard deviation divided by the mean of the yearly average high pulse durations.
\item dh17 High flow duration. Compute the average duration of flow events with flows above a threshold equal to 
the median flow value for the entire flow record. dh17 is the mean duration 
of the events.
\item dh18 High flow duration. Compute the average duration of flow events with flows above a threshold equal to 
three times the median flow value for the entire flow record. dh18 is the mean 
duration of the events.
\item dh19 High flow duration. Compute the average duration of flow events with flows above a threshold equal to 
seven times the median flow value for the entire flow record. dh19 is the mean
duration of the events .
\item dh20 High flow duration. Compute the 75th percentile value for the entire flow record. Compute the average 
duration of flow events with flows above a threshold equal to the 75th percentile value for the median annual 
flows. dh20 is the average duration of the events.
\item dh21 High flow duration. Compute the 25th percentile value for the entire flow record. Compute the average 
duration of flow events with flows above a threshold equal to the 25th percentile value for the entire set 
of flows. dh21 is the average duration of the events. 
\item dh22 Flood interval. Compute the flood threshold as the flow equivalent for a flood recurrence of 1.67 years. 
Determine the median number of days between flood events for each year. dh22 is the mean (or median-Use 
Preference option) of the yearly median number of days between flood events.
\item dh23 Flood duration. Compute the flood threshold as the flow equivalent for a flood recurrence of 1.67 years. 
Determine the number of days each year that the flow remains above the flood threshold. DH23 is the mean (or 
median-Use Preference option) of the number of flood days for years in which floods occur.
\item dh24 Flood-free days. Compute the flood threshold as the flow equivalent for a flood recurrence of 1.67 years. 
Compute the maximum number of days that the flow is below the threshold for each year. DH24 is the mean 
(or median-Use Preference option) of the maximum yearly no-flood days.
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
hitDurationHigh(x=x,yearType=yearType)

}
