% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitFrequencyLow.R
\name{hitFrequencyLow}
\alias{hitFrequencyLow}
\title{Indices describing frequency of low flow events.}
\usage{
hitFrequencyLow(x, yearType = "water", digits = 3, pref = "mean", ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 3 indices used to describe the frequency of low flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item FL1; Low flood pulse count. Compute the average number of flow events with flows below a threshold equal to the 
25th percentile value for the entire flow record. FL1 is the average (or median-Use Preference option) number of 
events.
\item FL2; Variability in low pulse count. Compute the standard deviation in the annual pulse counts for FL1. FL2 is 
100 times the standard deviation divided by the mean pulse count.  
\item FL3; Frequency of low pulse spells. Compute the average number of flow events with flows below a threshold 
equal to 5 percent of the mean flow value for the entire flow record. FL3 is the average (or median-Use 
Preference option) number of events.
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
hitFrequencyLow(x=x,yearType=yearType)
}
