% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitRateChange.R
\name{hitRateChange}
\alias{hitRateChange}
\title{Indices describing rate of change of flow.}
\usage{
hitRateChange(x, yearType = "water", digits = 3, pref = "mean", ...)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating whether to use water years or calendar years, respectively.}

\item{digits}{A numeric. Number of digits to round indice values}

\item{pref}{A character of either "mean" or "median", indicating whether to use mean or median. See details.}

\item{...}{Optional arguments needed for \code{hitAllStats} function}
}
\value{
A data.frame of flow statistics
}
\description{
Calculates 9 indices used to describe the rate of change of flow conditions. 
See Table X in the EflowStats package vignette for a full description of indices.
}
\details{
Descriptions of indices.
\itemize{
\item ra1; Rise rate. Compute the change in flow for days in which the change is positive for the entire flow record. 
RA1 is the mean (or median-Use Preference option) of these values.
\item ra2; Variability in rise rate. Compute the standard deviation for the positive flow changes. RA2 is 100 times 
the standard deviation divided by the mean.
\item ra3; Fall rate. Compute the change in flow for days in which the change is negative for the entire flow record. 
RA3 is the mean (or median-Use Preference option) of these values.
\item ra4; Variability in fall rate. Compute the standard deviation for the negative flow changes. RA4 is 100 times 
the standard deviation divided by the mean.
\item ra5; Number of day rises. Compute the number of days in which the flow is greater than the previous day. RA5 
is the number of positive gain days divided by the total number of days in the flow record.
\item ra6; Change of flow. Compute the log of the flows for the entire flow record. Compute the change in log of flow 
for days in which the change is positive for the entire flow record. RA6 is the median of these values.
\item ra7; Change of flow. Compute the log of the flows for the entire flow record. Compute the change in log of 
flow for days in which the change is negative for the entire flow record. RA7 is the median of these log 
values.
\item ra8; Number of reversals. Compute the number of days in each year when the change in flow from one day to the 
next changes direction. RA8 is the average (or median - Use Preference option) of the yearly values.
\item ra9; Variability in reversals. Compute the standard deviation for the yearly reversal values. RA9 is 100 times the 
standard deviation divided by the mean.
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
hitRateChange(x=x,yearType=yearType)
}
