% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ma41.45}
\alias{ma41.45}
\title{Function to return the MA41.45 hydrologic indicator statistics for a given data frame}
\usage{
ma41.45(qfiletemp, drain_area)
}
\arguments{
\item{qfiletemp}{data frame containing a "discharge" column containing daily flow values}

\item{drain_area}{numeric containing the drainage area of the selected site}
}
\value{
ma41.45 list of MA41-MA45 for the given data frame
}
\description{
This function accepts a data frame that contains columns named "discharge" and "year_val" and a value "drain_area"
containing the drainage area (in square miles) and calculates
MA41; Annual runoff. Compute the annual mean daily flows. MA41 is the mean of the annual means divided
by the drainage area (cubic feet per second/square mile-temporal).
MA42; Variability across annual flows.  MA42 is the maximum annual flow minus the minimum annual flow divided by
the median of mean annual flows (dimensionless-spatial).
MA43; Variability across annual flows.  Compute the first (25th percentile) and third (75th percentile) quartiles
and the 10th and 90th percentiles for the annual means (every year in the flow record). MA43 is the third quartile
minus the first quartile divided by the median of the annual means (dimensionless-spatial).
MA44; Variability across annual flows.  Compute the first (25th percentile) and third (75th percentile) quartiles
and the 10th and 90th percentiles for the annual means (every year in the flow record). MA44 is the 90th percentile
minus the 10th percentile divided by the median of the annual means (dimensionless-spatial).
MA45; Skewness in the annual flows. MA45 is the mean of the annual flow means minus the median of the annual means
divided by the median of the annual means (dimensionless-spatial).
}
\examples{
qfiletempf<-sampleData
drain_area<-56.5
ma41.45(qfiletempf,drain_area)
}

