% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medflowbyyear.R
\name{medflowbyyear}
\alias{medflowbyyear}
\title{Function to return the median flow by year for a given data frame}
\usage{
medflowbyyear(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
medflowbyyear numeric value of the median flow by year for the given data frame
}
\description{
This function accepts a data frame that contains columns named "discharge" and "year_val" and 
calculates the median flow by year for the entire record
}
\examples{
qfiletempf<-sampleData
medflowbyyear(qfiletempf)
}

