% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magHighFunctions.R
\name{mh14}
\alias{mh14}
\title{Function to return the MH14 hydrologic indicator statistic for a given data frame}
\usage{
mh14(x, ...)
}
\arguments{
\item{x}{data frame containing a "flow" column containing daily flow values}
}
\value{
mh14 numeric value of MH14 for the given data frame
}
\description{
This function accepts a data frame that contains columns named "flow", "year_val" and "month_val" and 
calculates MH14, median of annual maximum flows. Compute the annual maximum flows from monthly maximum flows. 
Compute the ratio of annual maximum flow to median annual flow for each year. MH14 is the median of these ratios 
(dimensionless-temporal).
}
\examples{
x<-sampleData
mh14(x)
}

