% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magHighFunctions.R
\name{mh18}
\alias{mh18}
\title{Function to return the MH18 hydrologic indicator statistic for a given data frame}
\usage{
mh18(x, ...)
}
\arguments{
\item{x}{data frame containing a "flow" column containing daily flow values}
}
\description{
This function accepts a data frame that contains columns named "flow" and "year_val" and 
calculates MH18, variability across annual maximum flows. Compute the logs (log10) of the maximum annual flows. 
Find the standard deviation and mean for these values. MH18 is the standard deviation times 100 divided by the 
mean (percent-spatial).
}
\examples{
x<-sampleData
mh18(x)
}

