% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magHighFunctions.R
\name{mh21.27}
\alias{mh21.27}
\title{Function to return the MH21-27 hydrologic indicator statistics for a given flow timeseries}
\usage{
mh21.27(x, ...)
}
\arguments{
\item{x}{data frame containing a "flow" column containing daily flow values}
}
\value{
A named numeric vector of MH21 through MH27 indicator statistics
}
\description{
This function accepts a vector of daily flow values and 
calculates the MH21 through MH27 statistics.
}
\examples{
x <- sampleData$flow
mh21.27(x)
}

