% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh24.R
\name{mh24}
\alias{mh24}
\title{Function to return the MH24 hydrologic indicator statistic for a given data frame}
\usage{
mh24(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
mh24 numeric value of MH24 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and 
calculates MH24, high peak flow. Compute the average peak flow value for flow events above a threshold equal 
to the median flow for the entire record. MH24 is the average peak flow divided by the median flow for the 
entire record (dimensionless-temporal).
}
\examples{
qfiletempf<-sampleData
mh24(qfiletempf)
}

