% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh25.R
\name{mh25}
\alias{mh25}
\title{Function to return the MH25 hydrologic indicator statistic for a given data frame}
\usage{
mh25(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
mh25 numeric value of MH25 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and 
calculates MH25, high peak flow. Compute the average peak-flow value for flow events above a threshold equal 
to three times the median flow for the entire record. MH25 is the average peak flow divided by the median flow 
for the entire record (dimensionless-temporal).
}
\examples{
qfiletempf<-sampleData
mh25(qfiletempf)
}

