% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ml18}
\alias{ml18}
\title{Function to return the ML18 hydrologic indicator statistic for a given data frame}
\usage{
ml18(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
ml18 numeric value of ML18 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and
calculates the variability in base flow index 1. Compute the standard deviation for the ratios of minimum 7-day
moving average flows to mean annual flows for each year.  ML18 is the standard deviation times 100 divided by
the mean of the ratios. (percent-spatial)
}
\examples{
qfiletempf<-sampleData
ml18(qfiletempf)
}

