% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ml21}
\alias{ml21}
\title{Function to return the ML21 hydrologic indicator statistic for a given data frame}
\usage{
ml21(x)
}
\arguments{
\item{x}{data frame containing a "discharge" column containing daily flow values}
}
\value{
ml21 numeric value of ML21 for the given data frame
}
\description{
This function accepts a data frame that contains columns named "discharge" and "year_val" and
calculates ML21, variability across annual minimum flows. Compute the mean and standard deviation for the
annual minimum flows. ML21 is the standard deviation times 100 divided by the mean (percent-spatial).
}
\examples{
qfiletempf<-sampleData
ml21(qfiletempf)
}

