% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ObservedStatsOtherMulti}
\alias{ObservedStatsOtherMulti}
\title{Function to return all flow statistics for a group of sites with local daily discharge datafiles}
\usage{
ObservedStatsOtherMulti(dataPath, stats, startDt = "", endDt = "",
  sepChar = ",")
}
\arguments{
\item{dataPath}{path to directory containing data files}

\item{stats}{string containing desired groups of statistics}

\item{startDt}{string containing start date if other than minimum in files}

\item{endDt}{string containing end date if other than maximum in files}

\item{sepChar}{string containing the datafile separator, default is comma}
}
\value{
statsout data frame containing requested statistics for each station
}
\description{
This function accepts a string containing the path to a directory containing csv files of daily discharge data
with columns containing site id, date (datetime in format 'YYYY-MM-DD') and discharge, a file of drainage areas,
and list of desired statistic groups and returns a data frame containing requested statistics
}
\examples{
dataPath <- "C:/Users/jlthomps/Documents/R/JData/modeled/"
stats="magnifSeven,magStat,flowStat,durStat,timStat,rateStat"
ObservedStatsOtherMulti(dataPath,stats)
}

