% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ObservedStatsUSGS}
\alias{ObservedStatsUSGS}
\title{Function to return all flow statistics for a set of USGS sites and start and end date}
\usage{
ObservedStatsUSGS(sites, startdate, enddate, stats)
}
\arguments{
\item{sites}{list of USGS station ids}

\item{startdate}{beginning water year, will be translated to 10/01}

\item{enddate}{ending water year, will be translated to 09/30}

\item{stats}{string containing desired groups of statistics}
}
\value{
statsout data frame containing requested statistics for each station
}
\description{
This function accepts a list of sites, start and end date and list of desired statistic groups and returns a data frame containing requested statistics
}
\examples{
sites <- c("02177000", "02178400")
startdate <- "2009"
enddate <- "2013"
stats="magnifSeven,magStat,flowStat,durStat,timStat,rateStat"
ObservedStatsUSGS(sites,startdate,enddate,stats)
}

