% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMonthlyMean}
\alias{plotMonthlyMean}
\title{Function to plot the monthly mean time series for a given daily data series}
\usage{
plotMonthlyMean(x, station)
}
\arguments{
\item{x}{data frame containing mean monthly values}

\item{station}{station number for plot title}
}
\description{
This function accepts a data frame containing daily data and a site number and returns a graph of mean monthly data
}
\examples{
qfiletempf<-sampleData
meanmonts<-monthlyMeanTs(qfiletempf)
plotMonthlyMean(meanmonts,'02178400')
}

