% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includedData.R
\docType{data}
\name{qfiletempf}
\alias{qfiletempf}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Jessica Thompson \email{jlthomps@usgs.gov}
}
\keyword{data}

