% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ra8.9.R
\name{ra8.9}
\alias{ra8.9}
\title{Function to return the RA8 and RA9 hydrologic indicator statistics for a given data frame}
\usage{
ra8.9(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
ra8.9 list containing RA8 and RA9 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates 
RA8; Number of reversals. Compute the number of days in each year when the change in flow from one day to the 
next changes direction. RA8 is the average (or median - Use Preference option) of the yearly values (days-temporal).
RA9; Variability in reversals. Compute the standard deviation for the yearly reversal values. RA9 is 100 times the 
standard deviation divided by the mean (percent-spatial).
}
\examples{
qfiletempf<-sampleData
ra8.9(qfiletempf)
}

