% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_EflowStats.R
\name{sampleData}
\alias{sampleData}
\title{Sample Streamflow data.frame}
\format{A data frame with 731 rows and 7 variables:
\describe{
  \item{wy_val}{integer of the water year}
  \item{date}{Date of the complete day}
  \item{discharge}{Discharge values}
  \item{month_val}{integer month value}
  \item{year_val}{integer year value}
  \item{day_val}{integer day value}
  \item{jul_val}{day of year}
}}
\description{
A dataset with a few years of complete water year based data.
}
