% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{validate_data Discharge timeseries screening}
\usage{
validate_data(x, yearType)
}
\arguments{
\item{x}{A dataframe containing a vector of date values in the first 
column and vector of numeric flow values in the second column.}

\item{yearType}{A charcter of either "water" or "calendar" indicating 
whether to use water years or calendar years, respectively.}
}
\value{
data.frame with rows sorted by date. Boolean FALSE if data is not
found to be valid. (See details)
}
\description{
Function to check dataframe inputs for appropriate data 
classes and screen for missing values.
}
\details{
Checks performed ensure the data is valid for use with other 
EflowStats functions. 
#' \enumerate{
  \item First column must be of class `Date`.
  \item Second must be of class `numeric`.
  \item `yearType` input must be either "water" or "calendar". 
  \item Every year as defined by the `yearType` input must be complete.
}
}
\examples{
x <- sampleData[c("date","discharge")]
yearType = "water"
validate_data(x=x,yearType=yearType)
}
