% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterYear.R
\name{waterYear}
\alias{waterYear}
\title{Function to return the water year for a given date}
\usage{
waterYear(x, numeric = TRUE)
}
\arguments{
\item{x}{A date vector}

\item{numeric}{Logical. Return values are numeric vector or vector of ordered factors}
}
\value{
A vector of numeric water years for each date if \code{numeric = TRUE} otherwise an ordered factor of water years if \code{numeric = FALSE}
}
\description{
Function to return the water year for a given date
}
\examples{
waterYear(sampleData$date)
}

