% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfi.R
\name{bfi}
\alias{bfi}
\title{Function to return the base flow index for a given data frame}
\usage{
bfi(x)
}
\arguments{
\item{x}{A numeric vector of consecutive daily mean discharge values}
}
\value{
bfi numeric value of the base flow index for the given data frame
}
\description{
This function accepts a vector of daily mean discharge values and 
calculates the base flow index of the daily flow values for the entire record
}
\examples{
x<-sampleData$discharge
bfi(x)
}

