% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHighEvents.R
\name{calcEvents}
\alias{calcEvents}
\title{Function to decompose a flow series into high flow events defined as flows above a given threshold.}
\usage{
calcEvents(x, threshold, type = "high")
}
\arguments{
\item{threshold}{Threshold value defining event}

\item{type}{character indicating type of event. High flags events above threshold, low flags events below threshold.}

\item{flow}{A vector of flow values}
}
\value{
A dataframe with columns "flow" and "event"
}
\description{
This function accepts a vector of flow values, tests if each value in the flow vector
is above or below a defined threshold, and classifies the flow vector into events. An event
is defined as consecutive entries above or below the given threshold.
}
\examples{
x <- sampleData$discharge
threshold <- median(x,na.rm=TRUE)
calcHighEvents(x,threshold)
}

