% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareStats.R
\name{CompareStats}
\alias{CompareStats}
\title{Function to return all flow statistics and comparison statistics for a set of USGS sites and start and end date}
\usage{
CompareStats(stats, sites = "", dataPath = "", startDt = "", endDt = "",
  sepChar = ",", dataPath2 = "", startDt2 = "", endDt2 = "")
}
\arguments{
\item{stats}{string containing desired groups of statistics}

\item{sites}{optional list of USGS station ids}

\item{dataPath}{path to local directory containing data files}

\item{startDt}{beginning water year, will be transalted to 10/01}

\item{endDt}{ending water year, will be translated to 09/30}

\item{sepChar}{string containing the datafile separator, default is comma}

\item{dataPath2}{path to local directory containing data files}

\item{startDt2}{beginning water year, will be transalted to 10/01}

\item{endDt2}{ending water year, will be translated to 09/30}
}
\value{
statsout data frame containing requested statistics for each station
}
\description{
This function accepts a list of sites, one or two data paths, one or two sets of start and end dates and list of desired statistic groups and returns a data frame containing requested statistics
}
\examples{
sites <- c("02177000", "02178400")
startdate <- "2003"
enddate <- "2008"
startdate2 <- "2009"
enddate2 <- "2013"
stats <- "magnifSeven,magStat,flowStat,durStat,timStat,rateStat"
\dontrun{
compareResults <- CompareStats(stats,sites=sites,
                               startDt=startdate,
                               endDt=enddate,
                               startDt2=startdate2,
                               endDt2=enddate2)
}                               
}

