% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dh4}
\alias{dh4}
\title{Function to return the DH4 hydrologic indicator statistic for a given data frame}
\usage{
dh4(qfiletempf, pref = "mean")
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}

\item{pref}{string containing a "mean" or "median" preference}
}
\value{
dh4 numeric containing DH4 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates
DH4; Annual maximum of 30-day moving average flows. Compute the maximum of 30-day moving average flows. Compute
the maximum of a 30-day moving average flow for each year. DH4 is the mean (or median-Use Preference option)
of these values (cubic feet per second-temporal).
}
\examples{
qfiletempf<-sampleData
dh4(qfiletempf)
}

