% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dl15.R
\name{dl15}
\alias{dl15}
\title{Function to return the DL15 hydrologic indicator statistic for a given data frame}
\usage{
dl15(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
dl15 numeric containing DL15 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates
DL15; Low exceedence flows. Compute the 90-percent exceedence value for the entire flow record. DL15 is the
exceedence value divided by the median for the entire record (dimensionless-spatial).
}
\examples{
qfiletempf<-sampleData
dl15(qfiletempf)
}

