% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDrainageArea.R
\name{getDrainageArea}
\alias{getDrainageArea}
\title{Function to return the drainage area of a site from the NWISWeb sites web service}
\usage{
getDrainageArea(site)
}
\arguments{
\item{site}{string containing the USGS site id}
}
\value{
drain_area a numeric value of the drainage area for a given site
}
\description{
This function accepts a url parameter that already contains the desired
NWIS site. It returns a drain_area value.
}
\examples{
sites<-"02177000"
\dontrun{
da <- getDrainageArea(sites)
}
}

