% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_magnif7.r
\name{magnifSeven}
\alias{magnifSeven}
\title{Function to return the magnificent seven statistics for a given data series
This is a function to compute the 7 statistics of daily streamflow
used by Archfield et al., under revision (June 2013). Input to the function is a
time series of streamflow with date in the format Y-m-d. Data should be arranged in 
four columns with names:  1) date, 2) discharge, 3) month_val and 4) year_val.}
\usage{
magnifSeven(timeseries1)
}
\arguments{
\item{timeseries1}{data frame of daily flow data}
}
\value{
magnif7 data frame of calculated statistics
}
\description{
Function to return the magnificent seven statistics for a given data series
This is a function to compute the 7 statistics of daily streamflow
used by Archfield et al., under revision (June 2013). Input to the function is a
time series of streamflow with date in the format Y-m-d. Data should be arranged in 
four columns with names:  1) date, 2) discharge, 3) month_val and 4) year_val.
}
\examples{
timeseries1<-sampleData
mgSeven <- magnifSeven(timeseries1)
}

