% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/th1.2.R
\name{th1.2}
\alias{th1.2}
\title{Function to return the TH1 and TH2 hydrologic indicator statistics for a given data frame}
\usage{
th1.2(qfiletempf)
}
\arguments{
\item{qfiletempf}{data frame containing a "discharge" column containing daily flow values}
}
\value{
th1.2 list containing TH1 and TH2 for the given data frame
}
\description{
This function accepts a data frame that contains a column named "discharge" and calculates
TH1; Julian date of annual maximum. Determine the Julian date that the maximum flow occurs for each year.
Transform the dates to relative values on a circular scale (radians or degrees). Compute the x and y components
for each year and average them across all years. Compute the mean angle as the arc tangent of y-mean divided by
x-mean. Transform the resultant angle back to Julian date (Julian day-spatial). and
TH2; Variability in Julian date of annual maxima. Compute the coefficient of variation for the mean x and y
components and convert to a date (Julian days-spatial).
}
\examples{
qfiletempf<-sampleData
th1.2(qfiletempf)
}

